add textField extension file and modify loginVC
