+++ MoneyAccounting/Extensions/TextField+Ext.swift
@@ -1,0 +1,44 @@
+//
+//  TextField+Ext.swift
+//  MoneyAccounting
+//
+//  Created by Tatiana Lazarenko on 3/7/24.
+//
+
+import UIKit
+
+extension UITextField {
+    func setLeftPaddingPoints(_ amount:CGFloat){
+        let paddingView = UIView(
+            frame: CGRect(
+                x: 0,
+                y: 0,
+                width: amount,
+                height: self.frame.size.height
+            )
+        )
+        self.leftView = paddingView
+        self.leftViewMode = .always
+    }
+    
+    func setRightPaddingPoints(_ amount:CGFloat) {
+        let paddingView = UIView(
+            frame: CGRect(
+                x: 0,
+                y: 0,
+                width: amount,
+                height: self.frame.size.height
+            )
+        )
+        self.rightView = paddingView
+        self.rightViewMode = .always
+    }
+    
+    func setCornerRadius() {
+        self.layer.cornerRadius = self.frame.height / 2
+    }
+    
+    func setBorderStyle() {
+        self.borderStyle = .none
+    }
+}
